ARROW-7654: [Python] Update CSV ConvertOptions Docs
Previously the documentation for ConvertOptions only mentions that a
mapping from column names to arrow types can be supplied via the
column_types argument. A pyarrow.Schema object may also be supplied from
which the mapping will be derived. The text added to the documentation
is specific that a mapping will be created as opposed to stating that
the pyarrow.Schema object will be used directly as some attributes of a
pyarrow.Field are not propagated to the table that will result from
using the ConvertOptions to read CSV data. For example, nullability is
not passed down to the schema of the new table.